Fixed unit tests and pulled out integration tests for ISONE (#130) (#151)
* #130 Added fake ISONE environment variables for unit tests.

This allows the class to be initialized and unit tests which don't make web
requests to pass.

* #130 Minor cleanup for read_fixtures(...) function.

Having the fixtures_base_path outside the function wasn't necessary
and was less clear.

* #130 Moved test fixture files into isone subdirectory.

* #130 Fixed mocked requests in ISONE unit tests.

Also I renamed some of the fixtures that were being used for expected
responses to be clearer.

* #130 Split ISONE integration tests out of unit tests.

* #130 Fixed up the LMP integration tests for ISONE.

Historical data from 2015 no longer exists. I changed the date to 2017
to kick the problem down the road for two years ;)

* #130 Fixed up mocked response encoding issue for Python 3.
