+++ tests/fixtures/isone/morningreport_current.json
