+++ tests/fixtures/isone/sevendayforecast_20160101.json
