+++ tests/fixtures/isone/sevendayforecast_current.json
